requirements.txt
# Core DL
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0

# Hugging Face
transformers==4.44.2
accelerate==0.33.0
datasets==3.0.1
evaluate==0.4.3
peft==0.11.1

# CUDA-only quantization (Linux x86_64 w/ CUDA)
bitsandbytes==0.43.3; platform_system=="Linux" and platform_machine=="x86_64"

# Retrieval / Embeddings
open_clip_torch==2.32.0
faiss-gpu==1.8.0; platform_system=="Linux" and platform_machine=="x86_64"
faiss-cpu==1.8.0; platform_system!="Linux" or platform_machine!="x86_64"

# Captioning / Evaluation
pycocotools==2.0.8
git+https://github.com/salaniz/pycocoevalcap.git

# Utilities
hydra-core==1.3.2
omegaconf==2.3.0
tqdm==4.66.5
rich==13.8.0
pandas==2.2.3
matplotlib==3.9.2
pillow==10.4.0
opencv-python==4.10.0.84

# Experiment tracking (optional)
wandb==0.17.8